Rename the component to `SQLPreviewSidebar`
