+++ frontend/src/metabase/query_builder/components/view/SQLPreviewSidebar/index.ts
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/no-default-export -- deprecated usage
-export { default } from "./ConvertQueryModal";
+export { default } from "./SQLPreviewSidebar";
