+++ frontend/src/metabase/query_builder/components/view/SQLPreviewSidebar/utils.ts
