+++ frontend/src/metabase/query_builder/components/view/View/QueryViewNotebook/QueryViewNotebook.jsx
@@ -2,7 +2,7 @@ import PropTypes from "prop-types";
 import { useEffect, useState } from "react";
 
 import Notebook from "metabase/query_builder/components/notebook/Notebook";
-import SQLPreviewSidebar from "metabase/query_builder/components/view/ConvertQueryModal";
+import SQLPreviewSidebar from "metabase/query_builder/components/view/SQLPreviewSidebar";
 
 import { NotebookContainer } from "./QueryViewNotebook.styled";
